# nsqlookupd: use map for producers - with benchmarks