I refactored and added a bit to @andyxning's benchmark from #1036 

Results:

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkLookupRegistrations8x8-4          1065          1088          +2.16%
BenchmarkLookupRegistrations8x64-4         2264          1165          -48.54%
BenchmarkLookupRegistrations64x64-4        18669         8364          -55.20%
BenchmarkLookupRegistrations64x512-4       134794        10933         -91.89%
BenchmarkLookupRegistrations512x512-4      1698650       150939        -91.11%
BenchmarkLookupRegistrations512x2048-4     5751659       177923        -96.91%
BenchmarkFindProducers8x8-4                244           619           +153.69%
BenchmarkFindProducers8x64-4               243           2259          +829.63%
BenchmarkFindProducers64x64-4              244           2437          +898.77%
BenchmarkFindProducers64x512-4             245           16447         +6613.06%
BenchmarkFindProducers512x512-4            277           16825         +5974.01%
BenchmarkFindProducers512x2048-4           277           83689         +30112.64%
```

take a look @mreiferson