@ploxiln 2018-06-02
After that, I realized: we usually look-up topics, not channels, which my benchmark was looking up producers for. And it looks like there's some "filtering" code that kicks in sometimes. I should probably benchmark looking up producers of topics.

So I split out topic-lookup benchmarks. The results were pretty much the same as for channel lookup.

But I realized that the typical `/lookup` operation of nsqlookups involved `FindRegistrations()`, and other stuff, so now I figure I should make a benchmark that more closely matches what `doLookup()` actually does ...