@ploxiln 2018-06-02
and the results:

```
$ benchcmp old.txt new.txt 
benchmark                                  old ns/op     new ns/op     delta
BenchmarkLookupRegistrations8x8-4          2969          2923          -1.55%
BenchmarkLookupRegistrations8x64-4         6334          3298          -47.93%
BenchmarkLookupRegistrations64x64-4        55507         30026         -45.91%
BenchmarkLookupRegistrations64x512-4       396369        39583         -90.01%
BenchmarkLookupRegistrations512x512-4      5135966       515863        -89.96%
BenchmarkLookupRegistrations512x2048-4     16354360      595718        -96.36%
BenchmarkDoLookup8x8-4                     1477          1885          +27.62%
BenchmarkDoLookup8x64-4                    1462          3584          +145.14%
BenchmarkDoLookup64x64-4                   6189          8770          +41.70%
BenchmarkDoLookup64x512-4                  6387          23065         +261.12%
BenchmarkDoLookup512x512-4                 45920         65297         +42.20%
BenchmarkDoLookup512x2048-4                45398         119196        +162.56%
```

It's probably a good idea to add benchmarks for typical lookup-protocol IDENTIFY and REGISTER operations (mostly `AddProducer()`). `LookupRegistrations()` is only called when a producer disconnects a tcp-lookup-protocol connection. (But it can still be pretty bad for it to lock the RegistrationDB for a long time.)