@ploxiln 2018-06-02
Results for the new REGISTER bench:

```
benchmark                                  old ns/op       new ns/op      delta
BenchmarkRegister8x8-4                     80615           79249          -1.69%
BenchmarkRegister8x64-4                    733022          741142         +1.11%
BenchmarkRegister64x64-4                   6035122         6548519        +8.51%
BenchmarkRegister64x512-4                  179271032       52564415       -70.68%
BenchmarkRegister512x512-4                 1505092668      416921240      -72.30%
BenchmarkRegister512x2048-4                17299945707     1770267744     -89.77%
```