@andyxning 2018-06-03
@ploxiln Thanks for the more detailed benchmark. The new scheme's decreased performance may be due to the map allocation and adjustment.

But the lookup is called periodically and the register is only called on the start of nsqd and new consumer connections.