@mreiferson 2018-06-03
I think I agree, since lookups can also happen in parallel.

Thanks for this — do you want to clean up these commits before we merge?